chore: more code deduplication
